pyproject.toml
[project]
name = "gestion-immo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "bidict>=0.23.1",
    "django>=5.2",
    "django-extensions>=4.1",
    "django-phonenumber-field>=8.1.0",
    "pandas>=2.2.3",
    "phonenumbers>=9.0.3",
    "pydot>=3.0.4",
    "pyparsing>=3.2.3",
]

[tool.ruff]
select = [
    "A",
    "B",
    "C4",
    "E",
    "F",
    "I",
    "PD",
    "NPY",
    "DJ",
    "INP",
    "TID",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    "BLE001",
    "FBT",
    "DTZ",
    "EM",
    "FA",
    "PERF",
    "PL",
    "RUF",
]
ignore = ["F401"]
fixable = ["ALL"]
unfixable = []


# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".github",
    ".hg",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "migrations",
    "node_modules",
    "static",
]

line-length = 100

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
